Added no-arg constructor for firestore. Fixed a lot of bugs. Recyclerview not showing all product images for now.

+++ app/src/main/java/com/teampym/onlineclothingshopapplication/data/models/Checkout.kt
@@ -9,5 +9,5 @@ import java.math.BigDecimal
 data class Checkout(
     val userId: String,
     val cart: List<Cart> = emptyList(),
-    val totalCost: BigDecimal
+    val totalCost: Double
 ) : Parcelable
